requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==9.5.0
python-jose==3.3.0
loguru==0.7.2
ddddocr==1.4.7
httpx==0.25.0
python-dotenv==1.0.0
opencv-python-headless==4.8.1.78
numpy==1.26.2